requirements.txt
aiogram>=3.4.1
python-dotenv>=1.0.1
tortoise-orm>=0.25.1
aiosqlite>=0.19.0
asyncpg>=0.29.0
better_profanity>=0.7.0
